+++ src/views/Dinners/Dinners.tsx
@@ -1,9 +1,9 @@
 import { Button } from "~/components/ui/button";
 import { Dinner } from "./Dinner";
-import { type DinnerWithTags } from "../utils/types";
+import { type DinnerWithTags } from "../../utils/types";
 import { AddDinner } from "./AddDinner";
 import { useState } from "react";
-import { cn } from "../lib/utils";
+import { cn } from "../../lib/utils";
 
 type Props = {
   dinners: DinnerWithTags[];
